compiling implementation type in the graph class
