+++ include/gl/graph_traits.hpp
@@ -1,10 +1,8 @@
 #pragma once
 
 #include "edge_descriptor.hpp"
-#include "edge_tags.hpp"
-#include "gl/impl/implementation_tags.hpp"
+#include "gl/impl/impl_tags_decl.hpp"
 #include "types/default_types.hpp"
-#include "types/type_traits.hpp"
 #include "vertex_descriptor.hpp"
 
 #include <memory>
