+++ include/gl/vertex_descriptor.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "edge_tags.hpp"
-#include "gl/impl/implementation_tags.hpp"
+#include "gl/impl/impl_tags_decl.hpp"
 #include "types/default_types.hpp"
 #include "types/type_traits.hpp"
 #include "types/types.hpp"
