+++ tests/source/test_adjacency_list.cpp
@@ -3,6 +3,7 @@
 #include "transforms.hpp"
 
 #include <gl/impl/adjacency_list.hpp>
+#include <gl/graph_traits.hpp>
 
 #include <doctest.h>
