+++ tests/source/test_adjacency_matrix.cpp
@@ -3,6 +3,7 @@
 #include "transforms.hpp"
 
 #include <gl/impl/adjacency_matrix.hpp>
+#include <gl/graph_traits.hpp>
 
 #include <doctest.h>
